Update conversant adapter for new prebid-server interface
